moving converts etc. to Adapter
Also fixed bug where column b and h were flipped
