+++ SConcrete_Adapter/Convert/ToBHoM/SectionProperty.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -19,10 +19,11 @@
  * You should have received a copy of the GNU Lesser General Public License     
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
- 
- using BH.Engine.Structure;
+
+using BH.Engine.Structure;
 using BH.oM.Adapters.SConcrete;
 using BH.oM.Geometry.ShapeProfiles;
+using BH.oM.Structure;
 using BH.oM.Structure.Elements;
 using BH.oM.Structure.SectionProperties;
 using System;
@@ -31,28 +32,50 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.Engine.Adapters.SConcrete
+namespace BH.Adapter.SConcrete
 {
-    public static partial class Query
+    public static partial class Convert
     {
         /***************************************************/
         /**** Public Methods                            ****/
         /***************************************************/
 
-        public static StructuralUsage1D GetUsage(this Bar bar)
+        public static ConcreteSection ToConcreteSection(this SConcreteConfig config, Dictionary<string, double> values)
         {
-            ConcreteSection section = (ConcreteSection)bar.SectionProperty;
-            if (bar.IsVertical())
+            ConcreteSection section = null;
+            //Create BHoM Sections
+
+            // TO DO: Call constructors with the other values
+            // TO DO: Check that dictionary has the keys needed
+            // TO DO: Implement units scaling
+
+            switch (config.MemberType)
             {
-                return StructuralUsage1D.Column;
-            }
-            else
-            {
-                return StructuralUsage1D.Beam;
+                case MemberType.CircColumn:
+                    if (values.ContainsKey("Cm D"))
+                    {
+                        section = Create.ConcreteCircularSection(values["Cm D"].FromUnit(config.Units, UnitType.Length));
+                    }
+                    break;
+                case MemberType.RectColumn:
+                case MemberType.LBeam:
+                case MemberType.RectBeam:
+                case MemberType.TBeam:
+                case MemberType.CShapeWall:
+                case MemberType.IShapeWall:
+                case MemberType.LShapeWall:
+                case MemberType.TShapeWall:
+                default:
+                    Engine.Reflection.Compute.RecordWarning($"Pull of Member Type {config.MemberType.ToString()} not implemented. A dummy section has been returned.");
+                    section = Create.ConcreteCircularSection(1);
+                    break;
             }
 
+            return section;
         }
+
 
         /***************************************************/
     }
+
 }
