+++ SConcrete_Adapter/Convert/ToSConcrete/Panel.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -20,44 +20,43 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using BH.oM.Structure.Results;
+using BH.oM.Adapters.SConcrete;
+using BH.oM.Structure.Elements;
+using BH.oM.Structure.SurfaceProperties;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using BH.oM.Adapters.SConcrete;
 
-namespace BH.Engine.Adapters.SConcrete
+namespace BH.Adapter.SConcrete
 {
     public static partial class Convert
     {
         /***************************************************/
-        /**** Public Methods                            ****/
+        /**** Private Methods                           ****/
         /***************************************************/
 
-        public static string ToSConcrete(this BarForce barForce, int i, SConcreteConfig config)
+        public static string ToSConcretePanelInformation(Panel panel, int i)
         {
+            string l = "6000";
+            string t = "400";
+            string x0 = "0";
+            string y0 = "3000";
+            string angle = "0";
+            string zonenoa = "1";
+            string zonenob = "2";
+            string vertd = "4";
+            string horzd = "4";
+            string curt = "1";
+            string verts = "400";
+            string horzs = "400";
+            string hooka = "0";
+            string hookb = "0";
 
-            string LC = i.ToString("F0");
-            string Nf = (barForce.FX.ToUnit(config.Units, UnitType.Force)).ToString("G");
-            string Tf = (barForce.MX.ToUnit(config.Units, UnitType.Moment)).ToString("G");
-            string Vfz = (barForce.FZ.ToUnit(config.Units, UnitType.Force)).ToString("G");
-            string Mfy = (barForce.MY.ToUnit(config.Units, UnitType.Moment)).ToString("G");
-            string Cmy = " 1";
-            string Vfy = (barForce.FY.ToUnit(config.Units, UnitType.Force)).ToString("G");
-            string Mfz = (barForce.MZ.ToUnit(config.Units, UnitType.Moment)).ToString("G");
-            string Cmz = " 1";
-            string Pdistr = " 0";
-            string CheckLC = " True";
-            string LoadType = " 1";
-            string Comment = "--";
-            string AutoGen = "False";
-
-            return $"{LC}	{Nf}	{Tf}	{Vfz}	{Mfy}	{Cmy}	{Vfy}	{Mfz}	{Cmz}	{Pdistr}	{CheckLC}	{LoadType}	{Comment}	{AutoGen}" + Environment.NewLine;
+            return $" {i}	 True	1	 {l}	 {t}	 {x0}    {y0}    {angle}	 {zonenoa}	 {zonenob}	 {vertd}	 {horzd}	 {curt}	 {verts}	 {horzs}	 {hooka}	 {hookb}";
         }
 
         /***************************************************/
-
     }
 }
